Move container creation logic into libkpod
Retain the Kubernetes CRI interface for now, pending further
refactoring.
